安全修复

+++ application/common/model/Annex.php
@@ -80,6 +80,9 @@ class Annex extends Base {
 
         $path = './';
         foreach($list['list'] as $k=>$v){
+            if (stripos($v['annex_file'], '../') !== false) {
+                continue;
+            }
             $pic = $path.$v['annex_file'];
             if(file_exists($pic) && (substr($pic,0,8) == "./upload") || count( explode("./",$pic) ) ==1){
                 unlink($pic);
